requirements.txt
flask==2.3.3
opencv-python==4.8.1.78
numpy==1.24.3
easyocr==1.7.0
Pillow==10.0.1
torch==2.0.1
torchvision==0.15.2